+++ src/attribute/attribute.rs
@@ -30,17 +30,6 @@ pub struct Attribute {
     pub name: AttributeQName,
     /// Attribute value.
     pub value: AttributeValue,
-    /// Visibility.
-    ///
-    /// Unlike many other DOM implementations, svgdom supports hiding of the attributes,
-    /// instead of removing them. Invisible attributes act just like other attributes,
-    /// but they will not be printed during SVG writing.
-    /// Unless you enable them via [`WriteOptions`].
-    ///
-    /// All attributes are visible by default.
-    ///
-    /// [`WriteOptions`]: struct.WriteOptions.html
-    pub visible: bool,
 }
 
 // TODO: fix docs
@@ -64,7 +53,6 @@ impl Attribute {
         Attribute {
             name: AttributeQNameRef::from(name).into(),
             value: AttributeValue::from(value),
-            visible: true,
         }
     }